Rename write to catalog, simplify the writing flow
